Moves `XCTestCase` snapshot helpers to separate file
